Stable Release Candidate
- A lot of experimental code related to founding bases has been torn
out. It wasn't working and I got fed up trying to make it work. It
wasn't fun anymore.
- In a career game, recovery is now based on the nearest open base. This
took some serious persuasion of KSP to temporarily accept a launchsite
as the space centre. More often than not the recovery report given still
states the range from the KSC but it's actually the range from the
nearest open base. I haven't figured out how to fix that widget yet.
- Added a Base Boss app to the flight scene in career games. Right now
all it does is tell you the range to the nearest base and nearest open
base.