Stable Release (v0.6)
* The Launch Site Selector is now draggable in the VAB and SPH.
* Added Base Boss app to tracking station and map view. Use this to
toggle icons of open and closed bases in career mode, and to toggle
icons of rocketpads, helipads, runways and other launchsites in all game
modes.
* The Base Boss app in flight-scenes has some new features.
* In career mode, if you are landed within 2000m of a closed base, using
the Base Boss you can now open that base for half the funds it would
cost in the launch site selector of the VAB and SPH.
* When landed within 2000m of a base, the facilities at that base will
be listed in the Base Boss. Currently this is just for information and
this feature will be expanded on in a future version of KK.
* A simple ATC feature can be enabled and disabled in the Base Boss.
When enabled, you will recieve messages from open bases when you enter
their airspace or leave their airspace. These are accessed by the KSP
message widget on the right end of the KSP toolbar. This feature will be
expanded on in a future version of KK (hopefully with the addition of
audio!)